BigDecimal - added support for fractional hours

+++ src/main/java/pl/edu/agh/mwo/reporter/model/report/Report1.java
@@ -2,13 +2,14 @@ package pl.edu.agh.mwo.reporter.model.report;
 
 import pl.edu.agh.mwo.reporter.model.Person;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
 public class Report1 {
     private final String description = "Raport liczby godzim poswieconych na projekty w rozbiciu na pracownikow";
-    private final Map<Person, Integer> hoursPerPerson;
+    private final Map<Person, BigDecimal> hoursPerPerson;
 
     public Report1() {
         this.hoursPerPerson = new HashMap<>();
@@ -18,7 +19,7 @@ public class Report1 {
         return description;
     }
 
-    public Map<Person, Integer> getHoursPerPerson() {
+    public Map<Person, BigDecimal> getHoursPerPerson() {
         return hoursPerPerson;
     }
 
@@ -26,11 +27,11 @@ public class Report1 {
         return hoursPerPerson.keySet();
     }
 
-    public Integer getHoursForPerson(Person person) {
+    public BigDecimal getHoursForPerson(Person person) {
         return this.hoursPerPerson.get(person);
     }
 
-    public void addPersonWithTotalNumberOfHours(Person person, Integer totalNumberOfHours) {
+    public void addPersonWithTotalNumberOfHours(Person person, BigDecimal totalNumberOfHours) {
         this.hoursPerPerson.put(person, totalNumberOfHours);
     }
 }
